+++ src/main/java/com/example/Waffle/repository/UserGroupRepository.java
@@ -1,7 +1,7 @@
 package com.example.Waffle.repository;
 
-import com.example.Waffle.entity.UserGroup.UserGroupEntity;
-import com.example.Waffle.entity.UserGroup.UserGroupPK;
+import com.example.Waffle.entity.Group.UserGroupEntity;
+import com.example.Waffle.entity.Group.UserGroupPK;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -15,5 +15,5 @@ public interface UserGroupRepository extends JpaRepository<UserGroupEntity, User
 
     Optional<UserGroupEntity> findByUserIdAndGroupId(Long userId, Long groupId);
 
-    Optional<UserGroupEntity> deleteByGroupId(int groupId);
+    Optional<UserGroupEntity> deleteByGroupId(Long groupId);
 }
